Finishing cruds

+++ 19_implementacion1/AppConsultorioMySQL/src/main/java/com/meli/consultorio/services/dentist/DentistService.java
@@ -1,5 +1,6 @@
 package com.meli.consultorio.services.dentist;
 
+import com.meli.consultorio.exceptions.DentistNotFoundException;
 import com.meli.consultorio.models.Dentist;
 import com.meli.consultorio.models.dtos.DentistDTO;
 import com.meli.consultorio.repositories.IDentistRepository;
@@ -8,7 +9,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -25,23 +26,31 @@ public class DentistService implements IDentistService{
 
     @Override
     public void deleteDentistById(Long id) {
+        if(iDentistRepository.findById(id).isEmpty()) {
+            throw new DentistNotFoundException(id);
+        }
+
         iDentistRepository.deleteById(id);
     }
 
     @Override
     public void updateDentist(DentistDTO dentist) {
+        if(iDentistRepository.findById(dentist.getId()).isEmpty()) {
+            throw new DentistNotFoundException(dentist.getId());
+        }
+
         iDentistRepository.save(mapper.map(dentist,Dentist.class));
     }
 
     @Override
-    public List<DentistDTO> findAllDentists() {
+    public Set<DentistDTO> findAllDentists() {
         List<Dentist> dentists = iDentistRepository.findAll();
-        return dentists.stream().map(dentist -> mapper.map(dentist, DentistDTO.class)).collect(Collectors.toList());
+        return dentists.stream().map(dentist -> mapper.map(dentist, DentistDTO.class)).collect(Collectors.toSet());
     }
 
     @Override
     public DentistDTO findDentistById(Long id) {
-        Optional<Dentist> item = iDentistRepository.findById(id);
-        return item.map(dentist -> mapper.map(dentist, DentistDTO.class)).orElse(null);
+        Dentist dentist = iDentistRepository.findById(id).orElseThrow(() -> new DentistNotFoundException(id));
+        return mapper.map(dentist, DentistDTO.class);
     }
 }
